PhpUse#getOriginal is deprecated, use #getFQN instead (part 2)
PhpUse#getOriginal is deprecated since PhpStorm 2016.1. This method was always returning almost the same value as #getFQN so this change most probably will not break anything. Unfortunately I've no chance to check it so it would be great if you'll perform some tests before or just after the merge. This pull request is rather notification than complete fix.